requirements.txt
asgiref==3.10.0
Django==5.2.7
django-debug-toolbar==6.0.0
django-filter==25.2
djangorestframework==3.16.1
drf-nested-routers==0.95.0
pillow==12.0.0
psycopg==3.2.11
psycopg-binary==3.2.11
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
